Move DeviceForm into SettingsForm
