+++ src/RocketBotGUI/App.config
@@ -87,11 +87,11 @@
     <add key="FirmwareType" value="user" />
     <add key="FirmwareFingerprint" value="OnePlus/OnePlus2/OnePlus2:6.0.1/MMB29M/1447840820:user/release-keys" />
     <!--Pokemon filter, seperated by comma-->
-    <add key="ExcludedPokemonCatch" value =""/>
+    <add key="ExcludedPokemonCatch" value="" />
     <!--Pokemon filter, seperated by comma-->
-    <add key="ExcludedPokemonTransfer" value =""/>
+    <add key="ExcludedPokemonTransfer" value="" />
     <!--Pokemon filter, seperated by comma-->
-    <add key="ExcludedPokemonEvolve" value =""/>
+    <add key="ExcludedPokemonEvolve" value="" />
   </appSettings>
   <system.web>
     <membership defaultProvider="ClientAuthenticationMembershipProvider">
